+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.12.4@DEV
   * If callable in `Dollection::getDocument()` specified, document always loaded directly omitting document pool.
+  * Fixed 'in' validator when custom error message specified
 
 ## 1.12.3 (2015-01-18)
   * Fixed bug with naming of validator in errors array. Removed suffix 'validator'.